NumberDialer: corrections for large numbers

+++ GLV/glv_textview.h
@@ -423,9 +423,9 @@ protected:
 	int numDigits() const { return mNI + mNF + numSignDigits(); }
 	int numSignDigits() const { return mShowSign ? 1:0; }
 	int signPos() const { return mShowSign ? 0 : -1; }
-	int valInt(int ix, int iy=0) const {
+	long long valInt(int ix, int iy=0) const {
 		double v = data().at<double>(ix,iy);
-		return (int)(v * pow(10., mNF) + (v>0. ? 0.5:-0.5));
+		return (long long)(v * pow(10., mNF) + (v>0. ? 0.5:-0.5));
 	}
 
 	double mag(int digit) const { return pow(10., numDigits()-1-digit - mNF); }
